requirements.txt
openai
dotenv
python-dotenv
requests
click
rich
tiktoken
pydantic
tomli
platformdirs
ddgs
httpx
beautifulsoup4
textual
pyinstaller
Pillow
typer   
